fix: made minor fixes to the code in order for it to be more tidy
