+++ package.json
@@ -46,7 +46,6 @@
     "@elastic/elasticsearch": "^8.13.1",
     "@godaddy/terminus": "^4.12.1",
     "@map-colonies/cleanup-registry": "^1.1.0",
-    "@map-colonies/detiler-common": "^1.0.0",
     "@map-colonies/error-express-handler": "^2.1.0",
     "@map-colonies/express-access-log-middleware": "^2.0.1",
     "@map-colonies/js-logger": "^1.0.1",
