+++ src/common/redis/interfaces.ts
@@ -1,0 +1,8 @@
+import { RedisClientOptions } from 'redis';
+
+export type RedisConfig = {
+    host: string;
+    port: number;
+    enableSslAuth: boolean;
+    sslPaths: { ca: string; cert: string; key: string };
+  } & RedisClientOptions;
